# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Naver Blog Assistant Local API
  version: 1.0.0
  description: >-
    Local web-app API for human-approved comment recommendations. The default listener is loopback;
    a separately paired device on trusted private Wi-Fi may use the same web app. Full article text
    is accepted only after preview confirmation and is not retained in full.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://127.0.0.1:8765
    description: Local machine only
...
security: []
paths:
...
  /api/v1/remote/pairing-code:
    post:
      tags: [Remote access]
      summary: Create Pairing Code
      operationId: createRemotePairingCode
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PairingCodeResponse"}
        "403":
          description: Only a local desktop can create pairing codes.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
            application/json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
        "409":
          description: Trusted-LAN mode must be enabled before a device can pair.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
            application/json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
  /api/v1/remote/pair:
    post:
      tags: [Remote access]
      summary: Pair Device
      operationId: pairRemoteDevice
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PairDeviceRequest"}
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PairDeviceResponse"}
        "403":
          description: The pairing code is invalid or expired.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
            application/json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
        "429":
          description: Pairing attempts are temporarily rate limited.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
            application/json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
        "422":
          description: Pairing request validation failed.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
            application/json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
  /api/v1/remote/devices:
    get:
      tags: [Remote access]
      summary: List Devices
      operationId: listRemoteDevices
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RemoteDeviceListResponse"}
        "403":
          description: Only a local desktop can list paired devices.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
            application/json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
  /api/v1/remote/devices/{session_id}:
    delete:
      tags: [Remote access]
      summary: Revoke Device
      operationId: revokeRemoteDevice
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Session Id
      responses:
        "204":
          description: The paired device was revoked.
        "403":
          description: Only a local desktop can revoke paired devices.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
            application/json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
        "404":
          description: The paired device does not exist or is already revoked.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
            application/json:
              schema: {$ref: "#/components/schemas/ProblemDetails"}
        "422":
          description: Validation Error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HTTPValidationError"}
components:
...
  schemas:
    PairingCodeResponse:
      properties:
        code: {type: string, title: Code}
        expires_at: {type: string, format: date-time, title: Expires At}
      additionalProperties: false
      type: object
      required: [code, expires_at]
      title: PairingCodeResponse
    PairDeviceRequest:
      properties:
        code: {type: string, maxLength: 128, minLength: 1, title: Code}
        device_name: {type: string, maxLength: 80, minLength: 1, title: Device Name}
      additionalProperties: false
      type: object
      required: [code, device_name]
      title: PairDeviceRequest
    RemoteDeviceResponse:
      properties:
        id: {type: string, format: uuid, title: Id}
        device_name: {type: string, maxLength: 80, minLength: 1, title: Device Name}
        created_at: {type: string, format: date-time, title: Created At}
        last_seen_at: {type: string, format: date-time, title: Last Seen At}
        expires_at: {type: string, format: date-time, title: Expires At}
      additionalProperties: false
      type: object
      required: [id, device_name, created_at, last_seen_at, expires_at]
      title: RemoteDeviceResponse
    PairDeviceResponse:
      properties:
        device: {$ref: "#/components/schemas/RemoteDeviceResponse"}
      additionalProperties: false
      type: object
      required: [device]
      title: PairDeviceResponse
    RemoteDeviceListResponse:
      properties:
        items:
          type: array
          maxItems: 100
          title: Items
          items: {$ref: "#/components/schemas/RemoteDeviceResponse"}
      additionalProperties: false
      type: object
      required: [items]
      title: RemoteDeviceListResponse
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
          minLength: 1
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
          maxLength: 1000
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*$"
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...